Restructure to run Dozent as a module for help command
